+++ src/primitives/mod.rs
@@ -4,6 +4,7 @@
 //! including request and response structures.
 
 mod copy_request;
+mod delete_request;
 mod list_response;
 mod model_details;
 mod model_summary;
@@ -12,6 +13,7 @@ mod running_model;
 mod version;
 
 pub use copy_request::CopyRequest;
+pub use delete_request::DeleteRequest;
 pub use list_response::ListResponse;
 pub use model_details::ModelDetails;
 pub use model_summary::ModelSummary;